refactor(BaseColorScheme) : reducing boilerplate code
